requirements.txt
mcp-server-fastmcp>=0.1.0
